Allow overriding the argoexec binary location
This introduces the ARGO_EXEC_PATH environment variable and when set,
the value is used as the path to copy to /var/run/argo/argoexec.

The specific usecase is to allow building one container image with a
CGO_ENABLED=1 binary of argoexec and one with CGO_ENABLED=0. When the
cluster must be FIPS compliant, we use the CGO_ENABLED=1 binary to use
FIPS approved C libraries by setting the ARGO_EXEC_PATH environment
variable. When the cluster does not require FIPS, we will just use the
CGO_ENABLED=0 binary which allows much more flexibility for the
container image used by the main container as the C libraries don't need
to match from the build environment.

Relates:
https://issues.redhat.com/browse/RHOAIENG-24702

Signed-off-by: mprahl <mprahl@users.noreply.github.com>